FINAL1.0-main/backend/requirements.txt
fastapi==0.115.5
uvicorn==0.34.0
python-multipart==0.0.20
PyPDF2==3.0.1
spacy==3.7.2
python-dotenv==1.0.0
mysql-connector-python==9.1.0
